Move tests util tests to own file
